Initial HTML file
Prompt:
I want to keep the features to an absolute bare-bones. No need for complex loading states, nonces, disallowing duplicate names, etc. The security here is "security through obscurity," since the stakes are very low. This is for fun a fun secret santa exchange!

With all that, let's getting started.