Tries to fix bugs, I realize we don't have a CSS file!
Prompt:
Sounds good.

OK, this definitely does not work and has several bugs. I'll describe them to you one at a time.

1. We show all the "pages" on the initial load. That is, I see headers for "Secret Santa Setup," "Secret Santa Admin Page," and "Secret Santa Pairing." The last "Pairing" section also shows placeholder text: "Hello, ! You will be giving a gift to ."

Fix this. Remember, these "pages" should only be displayed for these conditions:

- Setup: shown on initial load, when no params are present.
- Admin: when we have generated our pairings
- When an individual clicks one of their personalized links that shows their pairing.

Let's store which page we are on in our application state, under a top level key named "page." `page` will be either "setup", "admin", or "pairing".

Note that we'll have a default value of this as "setup", for the case when you load the page for the first time and you don't have any application state in the hash param yet.