Application state adjustments
Prompt:
Thanks, this was an important missing piece.

OK, on with the next bugs.

2. The individual links don't actually work. We can't have a hash param URL like `#individual-0` because our compressed application state immediately overrides this.

Let's rethink our application state.

Here is what I'll want to save. I'll use typescript types to describe the state shape, but just write out the code in plain Javascript and annotate things with JSDoc comments.

```
type ApplicationState =
	| {
			page?: 'setup';
	  }
	| {
			page: 'admin';
			disallowedPairs: Array<[string, string]>;
			pairings: Array<[string, string]>;
			participants: Array<string>;
	  }
	| {
			page: 'pairing';
			giver: string;
			receiver: string;
	  };
```

Update the application code accordingly. Note too that for the individual links, we'll just pass in a `page: 'pairing'` application state object, and that will cause the page to render that individual's pairing.