Fix links
Prompt:
No, this is wrong. I wasn't really checking your work, but I see the bug now. The links should work via an addEventListener. Instead, compute all the compressed applicationStates for the links, and then apply them as real href attributes so I can copy the links and send them in an email to all the participants.