Tries to make random pairings
Doesn't really have backtracking logic.

Prompt:
Thanks, that is better.

OK, next bug.

3. The pairings aren't random, the graph walk loops through the neighbors in order! We need better logic here to pick a neighbor at random, and after going through the graph, if we aren't able to find a Hamiltonian path, back track and keep backtracking if we've seen this path before.

Change generatePairings to add this new logic. Here is a description of this backtracing logic if it is helpful:

"Let’s figure out the algorithm. Our goal is to find a Hamiltonian cycle in the graph using brute-force. We start with a random node. This node has some connected nodes. We pick one of its neighbors at random and go there. This node has connected nodes too. We pick one of them which hasn’t been visited yet. And so on.

At some steps, we can find out that there are no available nodes to pick. It might happen if there are no more connections (because of constraints) or if all neighbor nodes have been visited already. This issue looks like the issue we had with the naive algorithm. With a graph, it’s more clear that we can do a step back and pick another random node. If we did a step back and there are still no nodes to pick, we do another step back.

Repeat until we visit all nodes and return to the first one. The cycle is complete"