Looks like this maybe works?
But it has a bug. The disallowsPairings only work in one direction. It
needs to work in both directions.

Prompt:
OK, now let's fix generatePairings.js.

Remember, the algorithm needs to:

- When starting at a random node, next pick a *random* neighbor it is connected to.
- Continue doing this. If we reach a point where there are no valid neighbors left, then *backtrack* to the last selected neighbor and try to pick a _different_ random neighbor we haven't already picked.
- If there are no valid neighbors after backtracking, keep backtracking until we can try a different path.
- If we make it back all the way to our starting now, then we know there isn't a valid Hamiltonian path in this graph, and we can exit and return null.

Please implement these changes. If you need to adjust the data structure for graph then feel free to do so.