Module
Prompt:
Great, this looks like it is working. Not the most efficient, but probably gets the job done.

OK, back to the app. When I load the webpage, now I get an error

"Uncaught SyntaxError: export declarations may only appear at top level of a module"

Do you need to change the HTML to load the <script> tag as a module?