Tries to fix bug, doesn't work
Prompt:
Thanks.

However, this still has a bug. `disallowedSet` isn't working. It is still possible if I have `disallowedPairs` of `["A", "B"]` for it to generate a pairing of `["B", "A"]`.

Let's do this to change it.

Rather than creating disallowedSet, create `disallowedMapOfSets`, where the keys is a participant, and the value is a `Set` of people that are disallowed.

Note that if the `disallowedPairs` argument was `[["A", "B"]]`, it should create a Map that looks like

```
// [["A", "B"]] yields ->
new Map([
  ["A", new Set(["B"])],
  ["B", new Set(["A"])],
])
```

As another example:
```
// [["A", "B"], ["A", "C"]] yields ->
new Map([
  ["A", new Set(["B", "C"])],
  ["B", new Set(["A"])],
  ["C", new Set(["A"])],
])
```

Make this logical change, and update all parts of `generatePairings` accordingly.